refactor: localize Playwright browser binary helpers
Move the product-shipped Playwright browser metadata record, platform mapping, layout validation, and resolver from common_utils into the serverless browser-PDF renderer module.

The helpers only serve PlaywrightPDFRenderer, so keeping them next to the render path avoids exporting browser-PDF-specific behavior from the general common utility module.

Update the focused tests to import the browser binary helpers from pdf_renderer and patch the renderer module platform lookup.

Verification: py_compile on changed Python files and git diff --check passed. uv-backed pre-commit/check/smoketest and focused pytest could not run in this environment because uv is unavailable and the existing venv has an unreadable pytest file.
